--- a/.benchmark.sh
+++ b/.benchmark.sh
@@ -3,5 +3,7 @@
 set -e +o pipefail
 
 if [ "$TRAVIS_PULL_REQUEST" != "false" ] ; then
-   cat .benchmark.log
+    curl -H "Authorization: token ${GITHUB_TOKEN}" -X POST \
+        --data-urlencode "{\"body\": \"$(cat .benchmark.log | sed "s/\"/'/g")\"}" \
+        "https://api.github.com/repos/${TRAVIS_REPO_SLUG}/issues/${TRAVIS_PULL_REQUEST}/comments"
 fi