--- a/.benchmark.sh
+++ b/.benchmark.sh
@@ -3,6 +3,8 @@
 set -e +o pipefail
 
 if [ "$TRAVIS_PULL_REQUEST" != "false" ] ; then
+    # Test Token Length
+    echo ${#GITHUB_TOKEN}
     curl -H "Authorization: token ${GITHUB_TOKEN}" -X POST \
         --data-urlencode "{\"body\": \"$(cat .benchmark.log | sed "s/\"/'/g")\"}" \
         "https://api.github.com/repos/${TRAVIS_REPO_SLUG}/issues/${TRAVIS_PULL_REQUEST}/comments"