--- a/.travis.yml
+++ b/.travis.yml
@@ -46,3 +46,4 @@ script:
 
 after_success:
   - bash <(curl -s https://codecov.io/bash)
+  - bash .benchmark.sh