--- a/Makefile
+++ b/Makefile
@@ -72,7 +72,7 @@ ifeq ($(TEST_TYPE),benchmark)
 	( cd request; go test -run=NONE -bench=. -benchmem=true -tags 'etcd' ./... ) >> old
 	( cd core; go test -run=NONE -bench=. -benchmem=true -tags 'etcd' ./... ) >> old
 	( cd coremain; go test -run=NONE -bench=. -benchmem=true -tags 'etcd' ./... ) >> old
-	if command -v benchcmp; then benchcmp old new ; fi
+	if command -v benchcmp; then benchcmp old new > .benchmark.log ; cat .benchmark.log ; fi
 	git checkout -
 endif