@yongtang 2018-09-22
benchmark                       old ns/op     new ns/op     delta
BenchmarkCacheResponse-2        3610          4176          +15.68%
BenchmarkFileLookupDNSSEC-2     44813         44172         -1.43%
BenchmarkFileParseInsert-2      776157        783933        +1.00%
BenchmarkFileLookup-2           6675          14987         +124.52%
BenchmarkCache-2                103           138           +33.98%
BenchmarkMinimalTTL-2           84.0          144           +71.43%
BenchmarkRequestDo-2            2.66          2.66          +0.00%
BenchmarkRequestSize-2          3.04          3.04          +0.00%
BenchmarkRequestScrub-2         5715          5744          +0.51%
BenchmarkCoreServeDNS-2         430           794           +84.65%

benchmark                       old allocs     new allocs     delta
BenchmarkCacheResponse-2        29             29             +0.00%
BenchmarkFileLookupDNSSEC-2     194            194            +0.00%
BenchmarkFileParseInsert-2      678            678            +0.00%
BenchmarkFileLookup-2           64             64             +0.00%
BenchmarkCache-2                0              0              +0.00%
BenchmarkMinimalTTL-2           0              0              +0.00%
BenchmarkRequestDo-2            0              0              +0.00%
BenchmarkRequestSize-2          0              0              +0.00%
BenchmarkRequestScrub-2         99             99             +0.00%
BenchmarkCoreServeDNS-2         5              5              +0.00%

benchmark                       old bytes     new bytes     delta
BenchmarkCacheResponse-2        512           512           +0.00%
BenchmarkFileLookupDNSSEC-2     6720          6720          +0.00%
BenchmarkFileParseInsert-2      116289        116290        +0.00%
BenchmarkFileLookup-2           2064          2064          +0.00%
BenchmarkCache-2                0             0             +0.00%
BenchmarkMinimalTTL-2           0             0             +0.00%
BenchmarkRequestDo-2            0             0             +0.00%
BenchmarkRequestSize-2          0             0             +0.00%
BenchmarkRequestScrub-2         5808          5808          +0.00%
BenchmarkCoreServeDNS-2         160           160           +0.00%