@yongtang 2018-09-22
@miekg Looks like it works, though I am using my own token to post the message. Do we have a coredns bot that can be used to post the message? The bot only needs a `public_repo` right in the setting for the token.