@yongtang 2018-09-22
There is a `coredns-cibot` https://github.com/coredns-cibot though I don't have access to this machine account. /cc @chrisohaver