@miekg 2018-09-22
[ Quoting <notifications@github.com> in "Re: [coredns/coredns] [WIP] Add ben..." ]
>There is a `coredns-cibot` https://github.com/coredns-cibot though I don't have access to this machine account. /cc @chrisohaver

I think we can just create another secret. Le'me check