@yongtang 2018-09-22
Yes multiple tokens are possible. The comment only need `public_repo` access. I would prefer to use a bot to post the comment. Otherwise it will looks like my comments are all over the places.