@miekg 2018-09-22
[ Quoting <notifications@github.com> in "Re: [coredns/coredns] [WIP] Add ben..." ]
>Yes multiple tokens are possible. The comment only need `public_repo` access. I would prefer to use a bot to post the comment. Otherwise it will looks like my comments are all over the places.
>

ah, yes. I see, it's *you* who's posting. Yes that's suboptimal, we have corbot 
running on drone (github.com/miekg/dreck), that could do this, but still it 
needs access to the benchmark text.

/Miek

--
Miek Gieben