@miekg 2018-09-23
[ Quoting <notifications@github.com> in "Re: [coredns/coredns] [WIP] Add ben..." ]
>@miekg Looks like it works, though I am using my own token to post the message. Do we have a coredns bot that can be used to post the message? The bot only needs a `public_repo` right in the setting for the token.

Just linking to the right build from the details would work as well. You need to 
need callback into github then. But not sure if you can do this.