@chrisohaver 2018-09-24
>  If you have the coredns-cibot credential, you can create a token with public_repo access, then in Travis CI you can replace GITHUB_TOKEN with the token you create. Then everything should work.

Done