@miekg 2018-09-25
Prolly need to put a warning on the CPU benchmarks:

"The CPU benchmarks are performed on Travis VMs and vary widly between runs, you 
can't trust them. The memory benchmarks are OK"


>benchmark                       old ns/op     new ns/op     delta
>BenchmarkCacheResponse-2        3851          4103          +6.54%
>BenchmarkFileLookupDNSSEC-2     44949         39587         -11.93%
>BenchmarkFileParseInsert-2      708788        799236        +12.76%
>BenchmarkFileLookup-2           6710          14836         +121.10%
>BenchmarkCache-2                103           132           +28.16%
>BenchmarkMinimalTTL-2           83.6          159           +90.19%
>BenchmarkRequestDo-2            2.65          2.65          +0.00%
>BenchmarkRequestSize-2          3.03          3.02          -0.33%
>BenchmarkRequestScrub-2         6180          5843          -5.45%
>BenchmarkCoreServeDNS-2         444           812           +82.88%
>
>benchmark                       old allocs     new allocs     delta
>BenchmarkCacheResponse-2        29             29             +0.00%
>BenchmarkFileLookupDNSSEC-2     194            194            +0.00%
>BenchmarkFileParseInsert-2      678            678            +0.00%
>BenchmarkFileLookup-2           64             64             +0.00%
>BenchmarkCache-2                0              0              +0.00%
>BenchmarkMinimalTTL-2           0              0              +0.00%
>BenchmarkRequestDo-2            0              0              +0.00%
>BenchmarkRequestSize-2          0              0              +0.00%
>BenchmarkRequestScrub-2         99             99             +0.00%
>BenchmarkCoreServeDNS-2         5              5              +0.00%
>
>benchmark                       old bytes     new bytes     delta
>BenchmarkCacheResponse-2        512           512           +0.00%
>BenchmarkFileLookupDNSSEC-2     6720          6720          +0.00%
>BenchmarkFileParseInsert-2      116289        116289        +0.00%
>BenchmarkFileLookup-2           2064          2064          +0.00%
>BenchmarkCache-2                0             0             +0.00%
>BenchmarkMinimalTTL-2           0             0             +0.00%
>BenchmarkRequestDo-2            0             0             +0.00%
>BenchmarkRequestSize-2          0             0             +0.00%
>BenchmarkRequestScrub-2         5808          5808          +0.00%
>BenchmarkCoreServeDNS-2         160           160           +0.00%
>
>-- 
>You are receiving this because you were mentioned.
>Reply to this email directly or view it on GitHub:
>https://github.com/coredns/coredns/pull/2123#issuecomment-424172956
/Miek

--
Miek Gieben