@chrisohaver 2018-09-25
> Prolly need to put a warning on the CPU benchmarks:

If we're looking for reproducible CPU benchmarks, we could run this as part of the coredns/ci jobs, which run the tests on a physical server.